move shell module up
